BelowZero - Splitting CraftDataHandler into partial class files - This separates Subnautica and BelowZero specific implementations - The common code remains in the old CraftDataHandler class
