+++ SMLHelper/Handlers/CraftDataHandlerSub.cs
@@ -1,62 +1,17 @@
-﻿namespace SMLHelper.V2.Interfaces
+﻿#if SUBNAUTICA
+namespace SMLHelper.V2.Handlers
 {
-#if BELOWZERO
-    using System.Collections.Generic;
-#endif
     using Crafting;
+    using Interfaces;
+    using Patchers;
 
     /// <summary>
-    /// A handler class for adding and editing crafted and some non-crafted items.
+    /// A handler class for adding and editing crafted items.
     /// </summary>
-    public interface ICraftDataHandler
+    public partial class CraftDataHandler : ICraftDataHandler
     {
-#if BELOWZERO
 
-        /// <summary>
-        /// <para>Allows you to edit RecipeData for TechTypes.</para>
-        /// <para>Can be used for existing TechTypes too.</para>
-        /// </summary>
-        /// <param name="techType">The TechType whose TechData you want to edit.</param>
-        /// <param name="techData">The TechData for that TechType.</param>
-        /// <seealso cref="RecipeData"/>
-        void SetTechData(TechType techType, RecipeData techData);
-
-        /// <summary>
-        /// <para>Allows you to edit JsonValues Directly for TechTypes.</para>
-        /// <para>Can be used for existing TechTypes too.</para>
-        /// </summary>
-        /// <param name="techType">The TechType whose TechData you want to edit.</param>
-        /// <param name="jsonValue">The JsonValue for that TechType.</param>
-        /// <seealso cref="TechData.defaults"/>
-        void SetTechData(TechType techType, JsonValue jsonValue);
-
-        /// <summary>
-        /// Safely accesses the crafting data from a modded item.<para/>
-        /// WARNING: This method is highly dependent on mod load order. 
-        /// Make sure your mod is loading after the mod whose TechData you are trying to access.
-        /// </summary>
-        /// <param name="techType">The TechType whose TechData you want to access.</param>
-        /// <returns>The JsonValue from the modded item if it exists; Otherwise, returns <c>null</c>.</returns>
-        JsonValue GetModdedTechData(TechType techType);
-
-        /// <summary>
-        /// <para>Allows you to add ingredients for a TechType crafting recipe.</para>
-        /// <para>Can be used for existing TechTypes too.</para>
-        /// </summary>
-        /// <param name="techType">The TechType whose ingredient list you want to edit.</param>
-        /// <param name="ingredients">The collection of Ingredients for that TechType.</param>
-        /// <seealso cref="Ingredient"/>
-        void AddIngredients(TechType techType, ICollection<Ingredient> ingredients);
-
-        /// <summary>
-        /// <para>Allows you to add linked items for a TechType crafting recipe.</para>
-        /// <para>Can be used for existing TechTypes too.</para>
-        /// </summary>
-        /// <param name="techType">The TechType whose ingredient list you want to edit.</param>
-        /// <param name="linkedItems">The collection of linked items for that TechType</param>
-        void AddLinkedItems(TechType techType, ICollection<TechType> linkedItems);
-
-#elif SUBNAUTICA
+        #region Subnautica Specific Static Methods
 
         /// <summary>
         /// <para>Allows you to edit recipes, i.e. TechData for TechTypes.</para>
@@ -65,7 +20,10 @@
         /// <param name="techType">The TechType whose TechData you want to edit.</param>
         /// <param name="techData">The TechData for that TechType.</param>
         /// <seealso cref="TechData"/>
-        void SetTechData(TechType techType, ITechData techData);
+        public static void SetTechData(TechType techType, ITechData techData)
+        {
+            Main.SetTechData(techType, techData);
+        }
 
         /// <summary>
         /// <para>Allows you to edit recipes, i.e. TechData for TechTypes.</para>
@@ -74,9 +32,38 @@
         /// <param name="techType">The TechType whose TechData you want to edit.</param>
         /// <param name="techData">The TechData for that TechType.</param>
         /// <seealso cref="TechData"/>
-        void SetTechData(TechType techType, TechData techData);
+        public static void SetTechData(TechType techType, TechData techData)
+        {
+            Main.SetTechData(techType, techData);
+        }
 
-#endif
+        #endregion
+
+        #region Subnautica specific implementations
+
+        /// <summary>
+        /// <para>Allows you to edit recipes, i.e. TechData for TechTypes.</para>
+        /// <para>Can be used for existing TechTypes too.</para>
+        /// </summary>
+        /// <param name="techType">The TechType whose TechData you want to edit.</param>
+        /// <param name="techData">The TechData for that TechType.</param>
+        /// <seealso cref="TechData"/>
+        void ICraftDataHandler.SetTechData(TechType techType, ITechData techData)
+        {
+            CraftDataPatcher.CustomTechData[techType] = techData;
+        }
+
+        /// <summary>
+        /// <para>Allows you to edit recipes, i.e. TechData for TechTypes.</para>
+        /// <para>Can be used for existing TechTypes too.</para>
+        /// </summary>
+        /// <param name="techType">The TechType whose TechData you want to edit.</param>
+        /// <param name="techData">The TechData for that TechType.</param>
+        /// <seealso cref="TechData"/>
+        void ICraftDataHandler.SetTechData(TechType techType, TechData techData)
+        {
+            CraftDataPatcher.CustomTechData[techType] = techData;
+        }
 
         /// <summary>
         /// <para>Allows you to edit EquipmentTypes for TechTypes.</para>
@@ -84,7 +71,10 @@
         /// </summary>
         /// <param name="techType">The TechType whose EqiupmentType you want to edit.</param>
         /// <param name="equipmentType">The EquipmentType for that TechType.</param>
-        void SetEquipmentType(TechType techType, EquipmentType equipmentType);
+        void ICraftDataHandler.SetEquipmentType(TechType techType, EquipmentType equipmentType)
+        {
+            CraftDataPatcher.CustomEquipmentTypes[techType] = equipmentType;
+        }
 
         /// <summary>
         /// <para>Allows you to edit QuickSlotType for TechTypes.</para>
@@ -92,7 +82,10 @@
         /// </summary>
         /// <param name="techType">The TechType whose QuickSlotType you want to edit.</param>
         /// <param name="slotType">The QuickSlotType for that TechType.</param>
-        void SetQuickSlotType(TechType techType, QuickSlotType slotType);
+        void ICraftDataHandler.SetQuickSlotType(TechType techType, QuickSlotType slotType)
+        {
+            CraftDataPatcher.CustomSlotTypes[techType] = slotType;
+        }
 
         /// <summary>
         /// <para>Allows you to edit harvest output, i.e. what TechType you get when you "harvest" a TechType.</para>
@@ -100,7 +93,10 @@
         /// </summary>
         /// <param name="techType">The TechType whose harvest output you want to edit.</param>
         /// <param name="harvestOutput">The harvest output for that TechType.</param>
-        void SetHarvestOutput(TechType techType, TechType harvestOutput);
+        void ICraftDataHandler.SetHarvestOutput(TechType techType, TechType harvestOutput)
+        {
+            CraftDataPatcher.CustomHarvestOutputList[techType] = harvestOutput;
+        }
 
         /// <summary>
         /// <para>Allows you to edit how TechTypes are harvested.</para>
@@ -108,7 +104,10 @@
         /// </summary>
         /// <param name="techType">The TechType whose HarvestType you want to edit.</param>
         /// <param name="harvestType">The HarvestType for that TechType.</param>
-        void SetHarvestType(TechType techType, HarvestType harvestType);
+        void ICraftDataHandler.SetHarvestType(TechType techType, HarvestType harvestType)
+        {
+            CraftDataPatcher.CustomHarvestTypeList[techType] = harvestType;
+        }
 
         /// <summary>
         /// <para>Allows you to edit how much additional slices/seeds are given upon last knife hit.</para>
@@ -116,7 +115,10 @@
         /// </summary>
         /// <param name="techType">The TechType whose final cut bonus you want to edit.</param>
         /// <param name="bonus">The number of additional slices/seeds you'll receive on last cut.</param>
-        void SetHarvestFinalCutBonus(TechType techType, int bonus);
+        void ICraftDataHandler.SetHarvestFinalCutBonus(TechType techType, int bonus)
+        {
+            CraftDataPatcher.CustomFinalCutBonusList[techType] = bonus;
+        }
 
         /// <summary>
         /// <para>Allows you to edit item sizes for TechTypes.</para>
@@ -124,7 +126,10 @@
         /// </summary>
         /// <param name="techType">The TechType whose item size you want to edit.</param>
         /// <param name="size">The item size for that TechType.</param>
-        void SetItemSize(TechType techType, Vector2int size);
+        void ICraftDataHandler.SetItemSize(TechType techType, Vector2int size)
+        {
+            CraftDataPatcher.CustomItemSizes[techType] = size;
+        }
 
         /// <summary>
         /// <para>Allows you to edit item sizes for TechTypes.</para>
@@ -133,7 +138,10 @@
         /// <param name="techType">The TechType whose item size you want to edit.</param>
         /// <param name="x">The width of the item</param>
         /// <param name="y">The height of the item</param>
-        void SetItemSize(TechType techType, int x, int y);
+        void ICraftDataHandler.SetItemSize(TechType techType, int x, int y)
+        {
+            CraftDataPatcher.CustomItemSizes[techType] = new Vector2int(x, y);
+        }
 
         /// <summary>
         /// <para>Allows you to edit crafting times for TechTypes.</para>
@@ -141,7 +149,10 @@
         /// </summary>
         /// <param name="techType">The TechType whose crafting time you want to edit.</param>
         /// <param name="time">The crafting time, in seconds, for that TechType.</param>
-        void SetCraftingTime(TechType techType, float time);
+        void ICraftDataHandler.SetCraftingTime(TechType techType, float time)
+        {
+            CraftDataPatcher.CustomCraftingTimes[techType] = time;
+        }
 
         /// <summary>
         /// <para>Allows you to edit the cooked creature list, i.e. associate the unedible TechType to the cooked TechType.</para>
@@ -149,7 +160,10 @@
         /// </summary>
         /// <param name="uncooked">The TechType whose cooked creature counterpart to edit.</param>
         /// <param name="cooked">The cooked creature counterpart for that TechType.</param>
-        void SetCookedVariant(TechType uncooked, TechType cooked);
+        void ICraftDataHandler.SetCookedVariant(TechType uncooked, TechType cooked)
+        {
+            CraftDataPatcher.CustomCookedCreatureList[uncooked] = cooked;
+        }
 
         /// <summary>
         /// <para>Allows you to edit inventory background colors for TechTypes.</para>
@@ -157,39 +171,21 @@
         /// <param name="techType">The TechType whose BackgroundType you want to edit.</param>
         /// <param name="backgroundColor">The background color for that TechType.</param>
         /// <seealso cref="CraftData.BackgroundType"/>
-        void SetBackgroundType(TechType techType, CraftData.BackgroundType backgroundColor);
+        void ICraftDataHandler.SetBackgroundType(TechType techType, CraftData.BackgroundType backgroundColor)
+        {
+            CraftDataPatcher.CustomBackgroundTypes[techType] = backgroundColor;
+        }
 
         /// <summary>
         /// Allows you to add items to the buildable list.
         /// </summary>
         /// <param name="techType">The TechType which you want to add to the buildable list.</param>
-        void AddBuildable(TechType techType);
+        void ICraftDataHandler.AddBuildable(TechType techType)
+        {
+            CraftDataPatcher.CustomBuildables.Add(techType);
+        }
 
-        /// <summary>
-        /// Allows you to add items to the game's internal grouping system.
-        /// Required if you want to make buildable items show up in the Habitat Builder.
-        /// </summary>
-        /// <param name="group">The TechGroup you want to add your TechType to.</param>
-        /// <param name="category">The TechCategory (in the TechGroup) you want to add your TechType to.</param>
-        /// <param name="techType">The TechType you want to add.</param>
-        void AddToGroup(TechGroup group, TechCategory category, TechType techType);
-
-        /// <summary>
-        /// Allows you to add items to the game's internal grouping system.
-        /// Required if you want to make buildable items show up in the Habitat Builder.
-        /// </summary>
-        /// <param name="group">The TechGroup you want to add your TechType to.</param>
-        /// <param name="category">The TechCategory (in the TechGroup) you want to add your TechType to.</param>
-        /// <param name="techType">The TechType you want to add.</param>
-        /// <param name="after">Added TechType will be added after this TechType, for sorting purposes.</param>
-        void AddToGroup(TechGroup group, TechCategory category, TechType techType, TechType after);
-
-        /// <summary>
-        /// Allows you to remove an existing TechType from the game's internal group system.
-        /// </summary>
-        /// <param name="group">The TechGroup in which the TechType is located.</param>
-        /// <param name="category">The TechCategory in which the TechType is located.</param>
-        /// <param name="techType">The TechType which you want to remove.</param>
-        void RemoveFromGroup(TechGroup group, TechCategory category, TechType techType);
+        #endregion
     }
 }
+#endif
